+++ CHANGELOG.md
@@ -5,6 +5,7 @@ To install or update LODA, please follow the [installation instructions](https:/
 ### Bugfixes
 
 * Fix handling of protected programs
+* Fix progress monitor for native mining (non-BOINC)
 
 ### Enhancements